Modify upgrade logic. Add convenience testing.
Switch to using a boolean field instead of setting weight = 0 for
whether upgrade cards can be displayed on level up.

TESTING: Added purely editor script UpgradeResetOnPlay to facilitate
testing when we repeatedly enter and exit play mode to reset
ScriptableObjects (upgrades).

Added upgrade PickupXPRadius -- has a max applyCount of 5.
Modified IncreaseMovementSpeed to have a max applyCount of 5.
Modified the existing attack upgrades AOEPunch and
RepulsiveField to use the same applyCount logic.

Minor fix to reference grabbing on StartManager.

Now the chain of methods called when applying an upgrade is :
1. UpgradeManager.ConfirmSelection
2. upgrade.ApplyUpgrade (automatically keeps track of applyCount)
--- For player attack upgrades:
3. PlayerScript.UnlockOrUpgradeAttack
4. attack.UpgradeAttack --> need to implement for each concrete autoattack